backend/tasks/celery_app.py
"""
Celery task definitions for background processing.

Note: For simplicity in this implementation, we use a synchronous approach
with background tasks. In production, you would configure Celery with Redis:

from celery import Celery
celery_app = Celery('tasks', broker=settings.REDIS_URL)

@celery_app.task
def analyse_incident_task(incident_id: str):
    asyncio.run(analyse_incident(incident_id))
"""

import asyncio
import logging
from concurrent.futures import ThreadPoolExecutor
from services.ai_engine import analyse_incident

logger = logging.getLogger(__name__)

# Thread pool for background tasks (simulating Celery)
executor = ThreadPoolExecutor(max_workers=4)


def run_async(coro):
    """Run async function in a new event loop."""
    loop = asyncio.new_event_loop()
    asyncio.set_event_loop(loop)
    try:
        return loop.run_until_complete(coro)
    finally:
        loop.close()


class AnalyseIncidentTask:
    """Simulated Celery task for incident analysis."""
    
    @staticmethod
    def delay(incident_id: str):
        """
        Queue the incident for analysis.
        In production, this would use Celery's delay() method.
        """
        logger.info(f"Queueing incident {incident_id} for analysis")
        executor.submit(run_async, analyse_incident(incident_id))
        return {"task_id": f"task-{incident_id}", "status": "queued"}


analyse_incident_task = AnalyseIncidentTask()
